+++ database/migrations/2026_04_17_154010_create_cms_pricing_rules_table.php
@@ -11,7 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('cms_pricing_rules', function (Blueprint $table) {
+        if (!Schema::hasTable('cms_pricing_rules')) {
+
+            Schema::create('cms_pricing_rules', function (Blueprint $table) {
             $table->id();
             $table->foreignId('company_id')->unique()->constrained('cms_companies')->onDelete('cascade');
             
@@ -35,6 +37,8 @@ return new class extends Migration
             
             $table->timestamps();
         });
+
+        }
     }
 
     /**
